Separate mediator-webapp as a separate module
