+++ modules/mediator-client/pom.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Title:        MEDIator Data Sharing Framework
+  ~ Description:  Data Sharing for Reproducible Scientific Research.
+  ~ Licence:      Apache License Version 2.0 - http://www.apache.org/licenses/
+  ~
+  ~ Copyright (c) 2018, Pradeeban Kathiravelu <pradeeban.kathiravelu@tecnico.ulisboa.pt>
+  -->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<parent>
+		<artifactId>modules</artifactId>
+		<groupId>edu.emory.bmi</groupId>
+		<version>1.0-SNAPSHOT</version>
+	</parent>
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>mediator-client</artifactId>
+	<name>mediator-client</name>
+
+
+</project>
