+++ modules/mediator-webapp/README.md
@@ -1,0 +1,16 @@
+
+Executing MEDIator Web Application
+---------
+Instead of running MEDIator as a stand-alone application with RESTful interfaces, you may run the MEDIator web 
+application, which is based on MEDIator and Embedded Tomcat.   
+   
+
+     sh modules/mediator-core/target/bin/webapp
+
+   
+   
+
+Go to http://localhost:<EMBEDDED_TOMCAT_PORT>/mediator/ using your browser to access the Web Application.
+
+By default, http://localhost/mediator/
+
