+++ modules/mediator-webapp/src/main/java/edu/emory/bmi/mediator/webapp/DataRetriever.java
@@ -12,7 +12,6 @@ import edu.emory.bmi.mediator.core.TciaInitializer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.io.FileNotFoundException;
 import org.apache.catalina.LifecycleException;
 
 /**
