+++ modules/mediator-webapp/src/main/java/edu/emory/bmi/mediator/webapp/constants/WebappConstants.java
@@ -6,12 +6,13 @@
  * Copyright (c) 2014, Pradeeban Kathiravelu <pradeeban.kathiravelu@tecnico.ulisboa.pt>
  */
 
-package edu.emory.bmi.mediator.constants;
+package edu.emory.bmi.mediator.webapp.constants;
 
 /**
  * Constants common to the data replication and synchronization tool.
  */
-public final class CommonConstants {
-    public static final int REST_PORT = 80;
-    public static final String FILE_SEPARATOR = "/"; //remove this
+public final class WebappConstants {
+	public static final int EMBEDDED_TOMCAT_PORT = 80;
+	public static final String WEB_APP_RELATIVE_PATH = "modules/mediator-webapp/src/main/resources/webapp/";
+	public static final String FILE_SEPARATOR = "/";
 }
