+++ modules/mediator-webapp/src/main/resources/webapp/delete.html
