+++ modules/mediator-webapp/src/main/resources/webapp/init.html
