+++ modules/mediator-webapp/src/main/resources/webapp/queryRetrieve.html
