+++ modules/mediator-webapp/src/main/resources/webapp/searchSeries.html
