+++ modules/mediator-webapp/src/main/resources/webapp/series.vm
