+++ modules/mediator-webapp/src/main/resources/webapp/studies.vm
